requirements.txt
numpy 
opencv-python 
PyOpenGL 
PyOpenGL_accelerate 
glfw
pillow
PyYAML
